Moving water to its own namespace
