+++ tests/Food/WaterTest.php
@@ -7,7 +7,7 @@ namespace Namelivia\Fitbit\Tests;
 use Carbon\Carbon;
 use Mockery;
 use Namelivia\Fitbit\Api\Fitbit;
-use Namelivia\Fitbit\Food\Water;
+use Namelivia\Fitbit\Food\Water\Water;
 
 class WaterTest extends TestCase
 {
